refactor(agents-api): refactor root render router + fix tests
